backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
psycopg2-binary>=2.9,<3.0
stripe>=7.0,<8.0
python-dotenv>=1.0,<2.0
